# swagger.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "FinVeda API",
                 "version":  "v1"
             },
    "paths":  {
...
                  "/api/v1/CapitalAccounts/investment":  {
                                                             "post":  {
                                                                          "tags":  [
                                                                                       "CapitalAccounts"
                                                                                   ],
                                                                          "requestBody":  {
                                                                                              "content":  {
                                                                                                              "application/json":  "@{schema=}",
                                                                                                              "text/json":  "@{schema=}",
                                                                                                              "application/*+json":  "@{schema=}"
                                                                                                          }
                                                                                          },
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "OK"
                                                                                                    }
                                                                                        }
                                                                      }
                                                         },
                  "/api/v1/CapitalAccounts/withdrawal":  {
                                                             "post":  {
                                                                          "tags":  [
                                                                                       "CapitalAccounts"
                                                                                   ],
                                                                          "requestBody":  {
                                                                                              "content":  {
                                                                                                              "application/json":  "@{schema=}",
                                                                                                              "text/json":  "@{schema=}",
                                                                                                              "application/*+json":  "@{schema=}"
                                                                                                          }
                                                                                          },
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "OK"
                                                                                                    }
                                                                                        }
                                                                      }
                                                         },
                  "/api/v1/CapitalAccounts/summary/{partnerId}":  {
                                                                      "get":  {
                                                                                  "tags":  [
                                                                                               "CapitalAccounts"
                                                                                           ],
                                                                                  "parameters":  [
                                                                                                     {
                                                                                                         "name":  "partnerId",
                                                                                                         "in":  "path",
                                                                                                         "required":  true,
                                                                                                         "schema":  "@{type=string; format=uuid}"
                                                                                                     }
                                                                                                 ],
                                                                                  "responses":  {
                                                                                                    "200":  {
                                                                                                                "description":  "OK",
                                                                                                                "content":  "@{text/plain=; application/json=; text/json=}"
                                                                                                            }
                                                                                                }
                                                                              }
                                                                  },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter \u0027Bearer\u0027 [space] and then your valid token.",
                                                              "scheme":  "Bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ]
}